\name{Exp37a}
\docType{data}
\alias{Exp37a}
\title{Digital PCR output file}
\description{
  This is a data file produced by a BioMark 48.770 machine made by Fluidigm Corporation for a digital PCR experiment. The observations start on row 13 after 3 rows of headers. Column 1 contains panel as well as x- and y-locations of the partition in a character string format. Column 3 contains sample information, which in this comprises a letter representing type and a code representing dilution. Column 6 contains the assay. Column 8 contains the quantification cycle (called Ct Value) data.
}
\usage{Exp37a}
\format{Data frame}
\source{LGC}
\references{
Devonshire, Alison S., et al. "Highly Reproducible Absolute Quantification of Mycobacterium tuberculosis Complex by Digital PCR." Analytical chemistry 87.7 (2015): 3706-3713.
}
\keyword{datasets}